% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int_cost.R
\name{cost}
\alias{cost}
\title{Cost function}
\usage{
cost(X, y, theta, lambda)
}
\arguments{
\item{X}{Design matrix X.}

\item{y}{Vector of observations.}

\item{theta}{Vector of starting values for coefficients of length \code{ncol(X)+1}.}

\item{lambda}{Regularisation parameter.}
}
\value{
Returns the cost of the current parameters theta.
}
\description{
\code{cost()} .
}
\details{
Internal function supplied to \code{optim} which calculates cost function
}
\examples{
library(dplyr)
theta <- rep(1,2)
lambda <- 0
x <- mpg$displ
y <- mpg$hwy

optim_out <- optim(
  par = theta,
fn = function(t) J(cbind(1, x), y, t, lambda),
gr = function(t) gR(cbind(1, x), y, t, lambda),
method = "BFGS"
)
}

