% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.vlrr.R
\name{predict.vlrr}
\alias{predict.vlrr}
\title{Predict method for the \code{vlrr} method.}
\usage{
\method{predict}{vlrr}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \code{vlrr} model object.}

\item{newdata}{If A list containing new training examples. See examples.}
}
\value{
Prediction based on the model object and newdata.
}
\description{
\code{predict.vlrr} Predict method for \code{vlrr} function.
}
\details{
Predict method for \code{vlrr} function.
}
\examples{
library(vlrr)
library(ggplot2) # to load mpg data

model <- vlrr(hwy ~ displ, data = mpg)

predict(model, newdata = list(hwy = mtcars$disp, displ = mtcars$disp * 0.0163))
}

