% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.summary.vlrr.R
\name{print.summary.vlrr}
\alias{print.summary.vlrr}
\title{Print method for summary method for the \code{vlrr} function.}
\usage{
\method{print}{summary.vlrr}(x, ...)
}
\arguments{
\item{x}{A \code{summary.vlrr} object.}

\item{newdata}{If A list containing new training examples. See examples.}
}
\value{
Prediction based on the model object and newdata.
}
\description{
\code{print.summary.vlrr} Print method for summary method for
  the \code{vlrr} function.
}
\details{
Print method method for summary method of \code{vlrr} function.
}
\examples{
library(vlrr)
library(ggplot2) # to load mpg data

model <- vlrr(hwy ~ displ, data = mpg)
summary(model)
}

