% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.vlrr.R
\name{print.vlrr}
\alias{print.vlrr}
\title{Print method for \code{vlrr} function.}
\usage{
\method{print}{vlrr}(x, ...)
}
\arguments{
\item{x}{A model fo class \code{vlrr}.}
}
\value{
A list of length three giving model call, final gradient, and coefficients
}
\description{
\code{print.vlrr} Provides print method for the \code{vlrr}
  class..
}
\details{
Simple method to reduce the size of output from the \code{vlrr}
  class..
}
\examples{
library(vlrr)

model <- vlrr(mpg ~ disp, data = mtcars)
print(model)
}

