% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.vlrr.R
\name{summary.vlrr}
\alias{summary.vlrr}
\title{Summarise a \code{vlrr} model}
\usage{
\method{summary}{vlrr}(object, ...)
}
\arguments{
\item{object}{A model object of class \code{vlrr}.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A list giving: convergence and message (from \code{optim} call),
  model formula call, regularisation parameter lambda, a summary of model
  coefficients, and a mean squared error of the fitted values and the
  original observations.
}
\description{
Summarise a \code{vlrr} model.
}
\details{
S3 method for prducing summaries of \code{vlrr} model objects.
}
\examples{
library(vlrr)
model <- vlrr(mpg ~ disp, data = mtcars)
summary(model)
}

