% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcov.vlrr.R
\name{vcov.vlrr}
\alias{vcov.vlrr}
\title{Return Variance-Covariance matrix of a vlrr object}
\usage{
\method{vcov}{vlrr}(object, ...)
}
\arguments{
\item{object}{A model object of class \code{vlrr}.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Returns the variance-covariance matrix of the main parameters of a \code{vlrr} object.
}
\details{
S3 method for returning the variance-covriance matrix of a \code{vlrr} model object.
}
\examples{
library(vlrr)
data(mpg, package="ggplot2")
model <- vlrr(mpg ~ disp, data = mtcars)
vcov(model)
}

