% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vlrr.formula.R
\name{vlrr.formula}
\alias{vlrr.formula}
\title{Use a formula to define \code{vlrr} model}
\usage{
\method{vlrr}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{Formula specifying model.}

\item{data}{List specifying location or vectors called in formula.}

\item{...}{Further arguments to or from other methods.}
}
\value{
Object of class formula
}
\description{
S3 formula method for \code{vlrr} class.
}
\details{
Give the details of the \code{code} here.
}
\examples{
library(vlrr)
model <- vlrr(mpg ~ disp, data = mtcars)
model
}

