% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vlrr.default.R
\name{vlrr}
\alias{vlrr}
\title{Vectorised linear regression with regularisation}
\usage{
vlrr(x, y, lambda, ...)
}
\arguments{
\item{y}{Vector of observations.}

\item{lambda}{Regularisation parameter.}

\item{X}{Design matrix x.}
}
\value{
Returns a list giving the original model call, the final gradient
  given by the last iteration of the \code{gradient} function call, and the
  final coefficients.
}
\description{
\code{vlrr} Main function behind the vlrr package.
}
\details{
Essential a wrapper function for the \code{optim} function..
}
\examples{
library(vlrr)
data(mpg, package="ggplot2")
model <- vlrr(mpg ~ disp, data = mtcars)
model
}

