% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vlrr_est.R
\name{vlrr_est}
\alias{vlrr_est}
\title{Vectorised linear regression with regularisation}
\usage{
vlrr_est(x, y, lambda)
}
\arguments{
\item{y}{Vector of observations.}

\item{lambda}{Regularisation parameter.}

\item{X}{Design matrix x.}
}
\value{
Returns a list giving the original model call, the final gradient
  given by the last iteration of the \code{gradient} function call, and the
  final coefficients.
}
\description{
\code{vlrr_est} Main function behind the vlrr package.
}
\details{
Essential a wrapper function for the \code{optim} function..
}
\examples{
library(dplyr)
}

