% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int_gradient.R
\name{gradient}
\alias{gradient}
\title{Gradient function}
\usage{
gradient(X, y, theta, lambda)
}
\arguments{
\item{X}{Design matrix X.}

\item{y}{Vector of observations.}

\item{theta}{Vector of starting values for coefficients of length \code{ncol(X)+1}.}

\item{lambda}{Regularisation parameter.}
}
\value{
Returns the gradient of the current parameters theta.
}
\description{
\code{gradient()} .
}
\details{
Internal function supplied to \code{optim} which calculates gradient.
}

